requirements.txt
feedparser==6.0.12
psycopg2==2.9.11
python-dotenv==1.2.1
PyYAML==6.0.3
sgmllib3k==1.0.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
